requirements.txt
click>=8.0.0
requests>=2.28.0
rich>=13.0.0
pyyaml>=6.0
jinja2>=3.1.0
watchdog>=2.1.0
psutil>=5.9.0
tabulate>=0.9.0
pydantic>=1.10.0
colorama>=0.4.6
tldextract>=3.4.0
dnspython>=2.2.0
httpx>=0.24.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
cryptography>=38.0.0
